add area, min/max Count for the xml Servlet

+++ OracleSpatialWeb/src/de/fraunhofer/iais/spatial/entity/FlickrArea.java
@@ -20,12 +20,12 @@ public class FlickrArea implements Serializable {
 	private long id;
 	private Radius radius;
 	private String name;
-	private float area;
+	private double area;
 	private Point2D center;
 	private List<Point2D> geom;
 	private boolean isCached;
-
 	private long totalCount;
+
 	private SortedMap<String, Integer> yearsCount = Maps.newTreeMap();
 	private SortedMap<String, Integer> monthsCount = Maps.newTreeMap();
 	private SortedMap<String, Integer> daysCount = Maps.newTreeMap();
@@ -66,11 +66,11 @@ public class FlickrArea implements Serializable {
 		this.geom = geom;
 	}
 
-	public float getArea() {
+	public double getArea() {
 		return area;
 	}
 
-	public void setArea(float area) {
+	public void setArea(double area) {
 		this.area = area;
 	}
